make edits based on pr comments, create configuration file, update variable/function names
