+++ src/configuration.js
@@ -1,0 +1,27 @@
+// user selects one of these options when first adding an item to a list
+export const radioButtonOptions = [
+  {
+    option: 'soon',
+    value: 7,
+    defaultChecked: true,
+  },
+  {
+    option: 'kind of soon',
+    value: 14,
+    defaultChecked: false,
+  },
+  {
+    option: 'not soon',
+    value: 30,
+    defaultChecked: false,
+  },
+];
+
+//TODO remove this from Navigatio component and replace with Tailwind classes
+export const navButtonStyles = {
+  position: 'fixed',
+  bottom: '0',
+  width: '100%',
+  textAlign: 'center',
+  backgroundColor: '#fff',
+};
